+++ source/source_io/CMakeLists.txt
@@ -17,6 +17,7 @@ list(APPEND objects
     module_energy/nscf_fermi_surf.cpp
     module_energy/write_eig_occ.cpp
     module_bessel/numerical_basis.cpp
+    module_bessel/numerical_basis_overlap.cpp
     module_bessel/numerical_basis_jyjy.cpp
     module_bessel/numerical_descriptor.cpp
     module_output/print_info.cpp
